Current progress without your changes to SessionManager

+++ DiabetesTracker/app/src/androidTest/java/com/cs428/dit/diabetestracker/DiagnosisActivityScoreTest.java
@@ -14,7 +14,7 @@ public class DiagnosisActivityScoreTest {
 
     @Test
     public void testMaleHealth() {
-        User user = new User(21.0, 72.0, 25, 123, true, false);
+        User user = new User((float)21.0, (float)72.0, 25, 123, true, false);
         int score = user.calcScore();
 
         assertEquals(user.getScoreBasedOnAge(), 4);
@@ -32,7 +32,7 @@ public class DiagnosisActivityScoreTest {
     @Test
     public void testFeMaleHealth() {
 
-        User user = new User(21.0, 72.0, 25, 123, false, false);
+        User user = new User((float)21.0, (float)72.0, 25, 123, false, false);
         int score = user.calcScore();
 
         assertEquals(user.getScoreBasedOnAge(), 4);
@@ -51,7 +51,7 @@ public class DiagnosisActivityScoreTest {
 
     @Test
     public void testBadHealth() {
-        User user = new User(30.0, 89.0, 50, 150, true, true);
+        User user = new User((float)30.0, (float)89.0, 50, 150, true, true);
         int score = user.calcScore();
 
         assertEquals(user.getScoreBasedOnAge(), 13);
